Renamed WeatherListViewController, removed logging, refactored VC extensions to separate files, reorganised folders
