+++ WeatherApp/WeatherApp/Core/API/Service/NetworkService.swift
@@ -16,7 +16,6 @@ class NetworkService {
                 guard let data = data else { return }
                 do {
                     let result = try self.decoder.decode(T.self, from: data)
-                    print(result)
                     completion(result)
                 } catch (let error)  {
                     print(error)
