+++ WeatherApp/WeatherApp/Features/WeatherList/UserInterface/WeatherList/WeatherListViewController.swift
@@ -1,12 +1,11 @@
 import UIKit
 
-class WeatherDetailsViewController: UIViewController {
+class WeatherListViewController: UIViewController {
     
     private let currentWeatherNetworkService = CurrentWeatherNetworkService()
-    private let cellHeight: CGFloat = 200.0
-    private var weather = [Weather]()
     private let weatherTitle = "Weather"
-    
+    let cellHeight: CGFloat = 200.0
+    var weather = [Weather]()
     let tableView = UITableView()
     var refreshControl = UIRefreshControl()
     
@@ -50,27 +49,4 @@ class WeatherDetailsViewController: UIViewController {
     
 }
 
-extension WeatherDetailsViewController: UITableViewDelegate {
-    func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-        return cellHeight
-    }
-    
-    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        tableView.deselectRow(at: indexPath, animated: true)
-    }
-}
 
-extension WeatherDetailsViewController:  UITableViewDataSource {
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return weather.count
-    }
-    
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCell(withIdentifier: WeatherInfoCell.identifier) as! WeatherInfoCell
-        let weatherforCell = weather[indexPath.row]
-        
-        cell.configure(model: weatherforCell)
-        return cell
-    }
-    
-}
